Refactor line into SVG module
