+++ lib/chart/svg.ex
@@ -69,7 +69,7 @@ defmodule Contex.SVG do
     ]
   end
 
-  def line(points, smoothed,  opts \\ []) do
+  def line(points, smoothed, opts \\ []) do
     attrs = opts_to_attrs(opts)
 
     path = path(points, smoothed)
@@ -96,6 +96,58 @@ defmodule Contex.SVG do
     end)
   end
 
+  defp path(points, true) do
+    # Use Catmull-Rom curve - see http://schepers.cc/getting-to-the-point
+    # First point stays as-is. Subsequent points are draw using SVG cubic-spline
+    # where control points are calculated as follows:
+    # - Take the immediately prior data point, the data point itself and the next two into
+    # an array of 4 points. Where this isn't possible (first & last) duplicate
+    # Apply Cardinal Spline to Cubic Bezier conversion matrix (this is with tension = 0.0)
+    #    0       1       0       0
+    #  -1/6      1      1/6      0
+    #    0      1/6      1     -1/6
+    #    0       0       1       0
+    # First control point is second result, second control point is third result, end point is last result
+
+    initial_window = {nil, nil, nil, nil}
+
+    {_, window, last_p, result} =
+      Enum.reduce(points, {:first, initial_window, nil, ""}, fn p,
+                                                                {step, window, last_p, result} ->
+        case step do
+          :first ->
+            {:second, {p, p, p, p}, p, []}
+
+          :second ->
+            {:rest, bump_window(window, p), p, ["M ", coord(last_p)]}
+
+          :rest ->
+            window = bump_window(window, p)
+            {cp1, cp2} = cardinal_spline_control_points(window)
+            {:rest, window, p, [result, " C " | [coord(cp1), coord(cp2), coord(last_p)]]}
+        end
+      end)
+
+    window = bump_window(window, last_p)
+    {cp1, cp2} = cardinal_spline_control_points(window)
+
+    [result, " C " | [coord(cp1), coord(cp2), coord(last_p)]]
+  end
+
+  defp bump_window({_p1, p2, p3, p4}, new_p), do: {p2, p3, p4, new_p}
+
+  @spline_tension 0.3
+  @factor (1.0 - @spline_tension) / 6.0
+  defp cardinal_spline_control_points({{x1, y1}, {x2, y2}, {x3, y3}, {x4, y4}}) do
+    cp1 = {x2 + @factor * (x3 - x1), y2 + @factor * (y3 - y1)}
+    cp2 = {x3 + @factor * (x2 - x4), y3 + @factor * (y2 - y4)}
+
+    {cp1, cp2}
+  end
+
+  defp coord({x, y}) do
+    ~s| #{x} #{y}|
+  end
 
   def opts_to_attrs(opts), do: opts_to_attrs(opts, [])
